requirements.txt
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
asgiref==3.8.1
certifi==2025.4.26
click==8.1.8
colorama==0.4.6
distro==1.9.0
dotenv==0.9.9
h11==0.16.0
htmltools==0.6.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.9.0
linkify-it-py==2.0.3
markdown-it-py==3.0.0
mdit-py-plugins==0.4.2
mdurl==0.1.2
narwhals==1.38.2
openai==1.78.0
orjson==3.10.18
packaging==25.0
pillow==11.2.1
prompt_toolkit==3.0.51
pydantic==2.11.4
pydantic_core==2.33.2
PyPDF2==3.0.1
pytesseract==0.3.13
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
questionary==2.1.0
setuptools==80.3.1
shiny==1.4.0
sniffio==1.3.1
starlette==0.46.2
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
uc-micro-py==1.0.3
uvicorn==0.34.2
watchfiles==1.0.5
wcwidth==0.2.13
websockets==15.0.1
wheel==0.45.1
